+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/e
 - [`Fix`] Block APIs will now pull the rest of the transactions for the block automatically
 - [`Fix`] Fix bytecode JSON parsing in transaction parsing
 - Add ConcClient a concurrent client for many operations
+- Add SimulateTransaction API for single-key authenticated transactions
 
 # v0.2.0 (6/10/2024)